complete tech interview
