+++ random-users/README.md
@@ -4,13 +4,13 @@ El objetivo de esta prueba técnica es crear una aplicación similar a la que se
 
 Los pasos a seguir:
 
-- [ ] Fetch 100 rows of data using the API.
+- [x] Fetch 100 rows of data using the API.
 - [x] Display the data in a table format, similar to the example.
 - [x] Provide the option to color rows as shown in the example.
 - [x] Allow the data to be sorted by country as demonstrated in the example.
 - [x] Enable the ability to delete a row as shown in the example.
 - [x] Implement a feature that allows the user to restore the initial state, meaning that all deleted rows will be recovered.
-- [] Handle any potential errors that may occur.
+- [x] Handle any potential errors that may occur.
 - [x] Implement a feature that allows the user to filter the data by country.
 - [x] Avoid sorting users again the data when the user is changing filter by country.
-- [] Sort by clicking on the column header.
+- [x] Sort by clicking on the column header.
