+++ random-users/src/models/User.ts
@@ -1,0 +1,15 @@
+export interface User {
+  id: string
+  name: string
+  lastname: string
+  photo: string
+  country: string
+}
+
+export const parseUser = (json: any): User => ({
+  id: json.id.value,
+  name: json.name.first,
+  lastname: json.name.last,
+  photo: json.picture.thumbnail,
+  country: json.location.country
+})
